답글 수정 버튼 클릭 시 에러나는 거 픽스해야함

+++ kmpetition/src/main/java/com/myspring/kmpetition/admin/service/AdminServiceImpl.java
@@ -131,11 +131,15 @@ public class AdminServiceImpl implements AdminService {
 		
 		ReplyVO reply = (ReplyVO) replyMap.get("reply");
 		List<UploadVO> deleteList = (ArrayList<UploadVO>) replyMap.get("delete");
-		List<UploadVO> replyUpload = (ArrayList<UploadVO>) replyMap.get("insert");
+		List<UploadVO> uploadList = (ArrayList<UploadVO>) replyMap.get("insert");
 		
 		dao.updateReply(reply);
-		dao.deleteReplyUpload(deleteList);
-		dao.insertReplyUpload(replyUpload);
+		if(deleteList !=null) {
+			dao.deleteReplyUpload(deleteList);
+		}
+		if(uploadList!=null) {
+			dao.insertReplyUpload(uploadList);
+		}
 	}
 	
 	public void removeReply(int articleNO) throws Exception{
